requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.6.4
exceptiongroup==1.2.0
frozenlist==1.4.1
gigachain==0.1.9
gigachain-community==0.0.24
gigachain-core==0.1.27
gigachat==0.1.17
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langsmith==0.1.11
loguru==0.7.0
marshmallow==3.21.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.27
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.0.4
yarl==1.9.4
